// ui/src/api/models/VideoStatus.ts
/* eslint-disable */
/**
 * Engawa
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const VideoStatus = {
    Pending: 'Pending',
    ObtainingMetadata: 'Obtaining Metadata',
    ObtainedMetadata: 'Obtained Metadata',
    Downloading: 'Downloading',
    PendingDownload: 'Pending Download',
    Failed: 'Failed',
    Deleted: 'Deleted',
    Complete: 'Complete',
    Excluded: 'Excluded',
    Filtered: 'Filtered',
    CopyrightStrike: 'Copyright Strike',
    Dmca: 'DMCA'
} as const;
export type VideoStatus = typeof VideoStatus[keyof typeof VideoStatus];


export function instanceOfVideoStatus(value: any): boolean {
    for (const key in VideoStatus) {
        if (Object.prototype.hasOwnProperty.call(VideoStatus, key)) {
            if (VideoStatus[key as keyof typeof VideoStatus] === value) {
                return true;
            }
        }
    }
    return false;
}

export function VideoStatusFromJSON(json: any): VideoStatus {
    return VideoStatusFromJSONTyped(json, false);
}

export function VideoStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): VideoStatus {
    return json as VideoStatus;
}

export function VideoStatusToJSON(value?: VideoStatus | null): any {
    return value as any;
}

export function VideoStatusToJSONTyped(value: any, ignoreDiscriminator: boolean): VideoStatus {
    return value as VideoStatus;
}